// src/frontend/typescript/app/models/post.ts
/*
 * Copyright 2019 Daniil <razornd> Razorenov
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *    http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */

import {User} from './user';
import {Comment} from './comment';
import {Owner} from './owner';

export interface PostKey {
  ownerId: number;
  postId: number;
}

export interface Post {
  id: PostKey;
  dateTime: Date;
  text: string;
  owner: Owner;
  from?: User;
  comments?: Comment[];
}